requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
geoalchemy2
geopandas
shapely
python-dotenv
python-multipart
passlib[bcrypt]
bcrypt==3.1.7
python-jose[cryptography]
alembic
email-validator